add component CaptureButton
